Add ruleset-specific legacy skin providers
This moves implementation of osu! skinnables to OsuLegacySkin.
