+++ osu.Game.Rulesets.Osu/UI/OsuPlayfield.cs
@@ -42,7 +42,7 @@ namespace osu.Game.Rulesets.Osu.UI
                 },
                 // Todo: This should not exist, but currently helps to reduce LOH allocations due to unbinding skin source events on judgement disposal
                 // Todo: Remove when hitobjects are properly pooled
-                new LocalSkinOverrideContainer(null)
+                new SkinProvidingContainer(null)
                 {
                     RelativeSizeAxes = Axes.Both,
                     Child = HitObjectContainer,
