+++ osu.Game/Skinning/LegacySkin.cs
@@ -9,7 +9,6 @@ using System.Threading.Tasks;
 using osu.Framework.Allocation;
 using osu.Framework.Audio;
 using osu.Framework.Audio.Sample;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Animations;
 using osu.Framework.Graphics.Containers;
@@ -20,8 +19,6 @@ using osu.Framework.Text;
 using osu.Game.Database;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
-using osu.Game.Rulesets.Objects.Drawables;
-using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Rulesets.UI;
 using osuTK;
 using osuTK.Graphics;
@@ -34,21 +31,12 @@ namespace osu.Game.Skinning
 
         protected IResourceStore<SampleChannel> Samples;
 
-        /// <summary>
-        /// On osu-stable, hitcircles have 5 pixels of transparent padding on each side to allow for shadows etc.
-        /// Their hittable area is 128px, but the actual circle portion is 118px.
-        /// We must account for some gameplay elements such as slider bodies, where this padding is not present.
-        /// </summary>
-        private const float legacy_circle_radius = 64 - 5;
-
         public LegacySkin(SkinInfo skin, IResourceStore<byte[]> storage, AudioManager audioManager)
             : this(skin, new LegacySkinResourceStore<SkinFileInfo>(skin, storage), audioManager, "skin.ini")
         {
             // defaults should only be applied for non-beatmap skins (which are parsed via this constructor).
             if (!Configuration.CustomColours.ContainsKey("SliderBall")) Configuration.CustomColours["SliderBall"] = new Color4(2, 170, 255, 255);
         }
-
-        private readonly bool hasHitCircle;
 
         protected LegacySkin(SkinInfo skin, IResourceStore<byte[]> storage, AudioManager audioManager, string filename)
             : base(skin)
@@ -62,14 +50,6 @@ namespace osu.Game.Skinning
 
             Samples = audioManager.GetSampleStore(storage);
             Textures = new TextureStore(new TextureLoaderStore(storage));
-
-            using (var testStream = storage.GetStream("hitcircle@2x") ?? storage.GetStream("hitcircle"))
-                hasHitCircle |= testStream != null;
-
-            if (hasHitCircle)
-            {
-                Configuration.SliderPathRadius = legacy_circle_radius;
-            }
         }
 
         protected override void Dispose(bool isDisposing)
@@ -91,30 +71,6 @@ namespace osu.Game.Skinning
                 case "Play/osu/cursor":
                     if (GetTexture("cursor") != null)
                         return new LegacyCursor();
-
-                    return null;
-
-                case "Play/osu/sliderball":
-                    var sliderBallContent = getAnimation("sliderb", true, true, "");
-
-                    if (sliderBallContent != null)
-                    {
-                        var size = sliderBallContent.Size;
-
-                        sliderBallContent.RelativeSizeAxes = Axes.Both;
-                        sliderBallContent.Size = Vector2.One;
-
-                        return new LegacySliderBall(sliderBallContent)
-                        {
-                            Size = size
-                        };
-                    }
-
-                    return null;
-
-                case "Play/osu/hitcircle":
-                    if (hasHitCircle)
-                        return new LegacyMainCirclePiece();
 
                     return null;
 
@@ -352,99 +308,6 @@ namespace osu.Game.Skinning
                         Origin = Anchor.Centre,
                     }
                 };
-            }
-        }
-
-        public class LegacySliderBall : CompositeDrawable
-        {
-            private readonly Drawable animationContent;
-
-            public LegacySliderBall(Drawable animationContent)
-            {
-                this.animationContent = animationContent;
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(ISkinSource skin, DrawableHitObject drawableObject)
-            {
-                animationContent.Colour = skin.GetValue<SkinConfiguration, Color4?>(s => s.CustomColours.ContainsKey("SliderBall") ? s.CustomColours["SliderBall"] : (Color4?)null) ?? Color4.White;
-
-                InternalChildren = new[]
-                {
-                    new Sprite
-                    {
-                        Texture = skin.GetTexture("sliderb-nd"),
-                        Colour = new Color4(5, 5, 5, 255),
-                    },
-                    animationContent,
-                    new Sprite
-                    {
-                        Texture = skin.GetTexture("sliderb-spec"),
-                        Blending = BlendingParameters.Additive,
-                    },
-                };
-            }
-        }
-
-        public class LegacyMainCirclePiece : CompositeDrawable
-        {
-            public LegacyMainCirclePiece()
-            {
-                Size = new Vector2(128);
-            }
-
-            private readonly IBindable<ArmedState> state = new Bindable<ArmedState>();
-
-            private readonly Bindable<Color4> accentColour = new Bindable<Color4>();
-
-            [BackgroundDependencyLoader]
-            private void load(DrawableHitObject drawableObject, ISkinSource skin)
-            {
-                Sprite hitCircleSprite;
-
-                InternalChildren = new Drawable[]
-                {
-                    hitCircleSprite = new Sprite
-                    {
-                        Texture = skin.GetTexture("hitcircle"),
-                        Colour = drawableObject.AccentColour.Value,
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                    },
-                    new SkinnableSpriteText("Play/osu/number-text", _ => new OsuSpriteText
-                    {
-                        Font = OsuFont.Numeric.With(size: 40),
-                        UseFullGlyphHeight = false,
-                    }, confineMode: ConfineMode.NoScaling)
-                    {
-                        Text = (((IHasComboInformation)drawableObject.HitObject).IndexInCurrentCombo + 1).ToString()
-                    },
-                    new Sprite
-                    {
-                        Texture = skin.GetTexture("hitcircleoverlay"),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                    }
-                };
-
-                state.BindTo(drawableObject.State);
-                state.BindValueChanged(updateState, true);
-
-                accentColour.BindTo(drawableObject.AccentColour);
-                accentColour.BindValueChanged(colour => hitCircleSprite.Colour = colour.NewValue, true);
-            }
-
-            private void updateState(ValueChangedEvent<ArmedState> state)
-            {
-                const double legacy_fade_duration = 240;
-
-                switch (state.NewValue)
-                {
-                    case ArmedState.Hit:
-                        this.FadeOut(legacy_fade_duration, Easing.Out);
-                        this.ScaleTo(1.4f, legacy_fade_duration, Easing.Out);
-                        break;
-                }
             }
         }
